requirements.txt
python-dotenv>=0.15.0,<1.0
pymongo>=3.11,<4.0